Json setting added

+++ Nur.APIService/Models/Products/ProductResultDTO.cs
@@ -9,9 +9,10 @@ public class ProductResultDTO
     public long Id { get; set; }
     public string Name { get; set; }
     public decimal Price { get; set; }
-    public long Quantity { get; set; }
+    public double Quantity { get; set; }
     public string Description { get; set; }
     public Unit Unit { get; set; }
+
     public ProductCategoryDTO Category { get; set; }
     public AttachmentResultDTO Attachment { get; set; }
 }
